Document sharing default settings (ConfigFile).
This addresses #32 and #38.

Issue #32 suggest a change in the `ConfigFile#config_for_env(hash)`
method in order to allow a separate entry to use for shared settings,
while issue #38 provides an implementation of an additional,
non-environment, key `default` which default values can be added to.

It seems like the existing method of sharing settings does not have any
significant disadvantages, and just needed to be explicitly documented.