Remove Sinatra::Decompile API
This extension is obsolete as Sinatra now provides access to a pattern object,
thanks to Mustermann, so that we no longer need to decompose Regexps.

If you're using this API directly, or via an extension that I'm unaware of...
Please let me know and we'll figure something out.

Its highly likely noone was using this extension directly, and possible that
only Namespace was the only consumer via internal usage.

So for now, I'm removing this extension until someone complains.